refactor: Api통신 응답결과 타입 수정 및 provideRetrofit함수 수정
