+++ app/src/main/java/com/wasingun/seller_lounge/data/repository/TrendComparisonRepository.kt
@@ -2,12 +2,13 @@ package com.wasingun.seller_lounge.data.repository
 
 import com.wasingun.seller_lounge.data.model.trendcomparison.KeywordRequest
 import com.wasingun.seller_lounge.data.model.trendcomparison.KeywordResponse
-import com.wasingun.seller_lounge.data.source.NaverApiClient
+import com.wasingun.seller_lounge.network.ApiResponse
+import com.wasingun.seller_lounge.network.NaverApiClient
 import javax.inject.Inject
 
 class TrendComparisonRepository @Inject constructor(private val naverApiClient: NaverApiClient) {
 
-    suspend fun requestComparisonResult(keywordRequest: KeywordRequest): KeywordResponse {
+    suspend fun requestComparisonResult(keywordRequest: KeywordRequest): ApiResponse<KeywordResponse> {
         return naverApiClient.getComparedKeywordData(keywordRequest)
     }
 }
